+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 7.11
  - Added hccapx support for hashcat 3.40 or higher
+ - Decrypted hashcat trophy files beautified
 
 ### 7.1
  - Added default choice for some yes/no question. Regexp improved